Create EmilyDoyleVigenereCipher
I know am trying to make a program that opens and then encrypts a file using Vigenere's cipher with a randomly generated key. Then get that file and decrypt it using the same method in reverse. I got the code together, before calling the file, and tested it with a message and key I created. It ran successfully but now I am trying to make the key random and have it call on a file to encrypt not a message I provide. I am unclear abt what file to go abt calling to be encrypted/decrypted as well as how to set it up to genterate a random key.